GODRIVER-2816 clean up the example
